tasks/payoff/app/requirements.txt
flask
pillow
flask_limiter
pymemcache>3,<5.0.0
